pyproject.toml
[project]
name = "image_project"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jcmullwh", email = "78822128+jcmullwh@users.noreply.github.com"},
]
dependencies = [
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "pillow>=12.0.0",
    "openai>=2.8.1",
    "ffmpeg>=1.4",
    "ai-backend @ git+https://github.com/jcmullwh/ai_backend.git@main",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pdm.scripts]
generate = "main.py"
test = "pytest"
manual-upscale = "python scripts/manual_upscale.py"
list-plans = "python scripts/list_prompt_catalog.py plans"
list-stages = "python scripts/list_prompt_catalog.py stages"
